Update alternateui.pas
Font size different on Windows compared to OSX and Linux Mate; added in a font ratio that is set at compile time to allow the scaling of font; so that they do not over-wrap the controls.
Corrected an issue where the install cross compiler button did not work, and also tidied up a few small issues, one of which was that after you have installed components; when you go back to add more components the list is NOW cleared; this should stop any attempt to double install a just installed component.
When AUI is activated, the form title is changed to show the version currently (1.0.4) of the alternateui.
Note bgrabitmap has an issue with osx, in that the debug windows will populate itself with various canvas and bitmap handle errors. To stop this on MAC you need to add an the compiler option  dBGRABITMAP_BGRAPIXEL to just the BGRABITMAP package when compiling for OSX.

This is the first time I have tried changing things on github, so hope this is all ok.